TC-374 Flytt upsert-funksjon fra TiltakRepositoryV2 til TiltakRepositoryV3.
